Update RichSmartID search in DeRichSystem to avoid type mis-matches    caused by updates to RichSmartID.
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@178321 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: f879627db41ef2382eb29263e88990cd703ec502

+++ LHCb/Det/RichDet/src/Lib/DeRichSystem.cpp
@@ -327,7 +327,7 @@ StatusCode DeRichSystem::fillMaps( const Rich::DetectorType rich )
       if ( ID.isValid() )
       {
         inacts.push_back( ID );
-        if ( std::find( softIDs.begin(), softIDs.end(), ID) == softIDs.end() )
+        if ( std::find( softIDs.begin(), softIDs.end(), (int)ID ) == softIDs.end() )
           warning() << "Invalid smartID in list of inactive PDs: " << *inpd << endmsg;
       }
       else
